ai-site-monitor/requirements.txt
beautifulsoup4>=4.12.0
crawl4ai>=0.4.0
playwright
requests>=2.31.0
urllib3>=2.0.0,<3.0.0
python-dotenv
google-genai
aiohttp
tqdm>=4.65.0
